requirements.txt
flask
ollama
groq
python-dotenv